// frontend/src/api/generated/models/readinessScoreSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type ReadinessScoreSource = typeof ReadinessScoreSource[keyof typeof ReadinessScoreSource];


export const ReadinessScoreSource = {
  Vendor: 'Vendor',
  Computed: 'Computed',
} as const;