// frontend/src/api/generated/models/readinessDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { HealthMetricDto } from './healthMetricDto';
import type { ReadinessComponentDto } from './readinessComponentDto';
import type { ReadinessScoreSource } from './readinessScoreSource';

export interface ReadinessDto {
  /** @nullable */
  date?: string | null;
  isToday?: boolean;
  /** @nullable */
  score?: number | null;
  scoreSource?: ReadinessScoreSource;
  /** @nullable */
  hrvRmssdMs?: number | null;
  /** @nullable */
  hrvBaselineMs?: number | null;
  /** @nullable */
  restingHeartRateBpm?: number | null;
  /** @nullable */
  restingHeartRateBaselineBpm?: number | null;
  /** @nullable */
  sleepDurationMinutes?: number | null;
  /** @nullable */
  sleepScore?: number | null;
  /** @nullable */
  ctl?: number | null;
  /** @nullable */
  atl?: number | null;
  /** @nullable */
  hasPainOrIllness?: boolean | null;
  /** @nullable */
  components?: ReadinessComponentDto[] | null;
  /** @nullable */
  healthStatus?: HealthMetricDto[] | null;
}